Update MailBinding.java
Mitigating CWE-79: Improper Restriction of Operations within a Function vulnerability in the populateMailMessage method.
Even though when the function receives unsupported charset and the exception will be logged, the malicious code would still be injected into the mail message. 

To mitigate this vulnerability, the populateMailMessage method should check the charset before calling the asEncodedInternetAddress method. If the charset is not supported, the method should throw a CamelMailException.